Emulation process now printed to standard output for debug.

+++ CPU.h
@@ -60,7 +60,7 @@ private:
   uint8_t sound_timer_;
   uint8_t drawflag_;
   uint8_t index_register_;
-  uint8_t program_counter_;
+  uint64_t program_counter_;
   uint16_t stack_ptr_;
   uint16_t opcode_;
